Policy Iteration (to be tested)
policy_iteration returns a map {state: action} and a map {state: value}
(useful for visualization)

policy_evaluation: returns a map {state: value}. Set default discount to
1 so that if not set the evaluation is undiscounted

I did not use the get_policy function because I think there's a
misunderstanding in its declaration: it shouldn't need a discount
parameter, since the discount is used in the computation of state
utility. The policy is instead formulated by simply acting greedily on
those utilities. Anyway, the declaration can be changed and the function
used by copying the code present in policy_iteration